refactor(vocab): create central index for vocabulary data
This change introduces a central `index.ts` file within the `src/data/vocab` directory to act as a single source of truth for all vocabulary data.

Previously, components might have needed to know the specific file structure of the vocabulary data. Now, all vocabulary is aggregated into an `allBooksVocab` object, and helper functions like `getVocabByBookId` are provided for common access patterns. This simplifies imports and makes the system more scalable for adding new books.
